+++ sscanf_macro/src/attribute/field.rs
@@ -14,6 +14,7 @@ pub enum FieldAttributeKind {
         ty: syn::Type,
         tries: bool,
     },
+    // TODO: add 'Optional' attribute
 }
 
 impl FromAttribute<attr::Field, &'_ syn::Type> for FieldAttributeKind {